requirements.txt
anthropic>=0.40.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-dotenv>=1.0.0
reportlab>=4.2.0
tqdm>=4.66.0
